collect/pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Edwin Kofler <edwin@kofler.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
Telethon = "^1.24.0"
python-dotenv = "^0.20.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
